requirements.txt
blis==0.4.1
catalogue==0.0.8
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cymem==2.0.3
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
requests==2.22.0
six==1.13.0
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
tqdm==4.40.0
urllib3==1.25.7
wasabi==0.4.0
Werkzeug==0.16.0
zipp==0.6.0
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-2.2.5/es_core_news_sm-2.2.5.tar.gz
